make individual list func into component
