+++ frontend/src/global_types.ts
@@ -274,3 +274,10 @@ export type ServiceWorkerTestOutput = {
   live: boolean,
   message: string,
 }
+
+export type FilterText = {
+  value: string;
+  onChange: React.Dispatch<React.SetStateAction<string>>;
+  disabled: boolean;
+  setDisabled: React.Dispatch<React.SetStateAction<boolean>>;
+}
