+++ frontend/src/pages/operation_list/individual_list/stylesheet.styl
@@ -1,0 +1,7 @@
+.operationList
+  display: flex
+  flex-wrap: wrap
+
+.opTitle
+  font-size: x-large
+  margin-bottom: 10px
