+++ frontend/src/pages/operation_list/stylesheet.styl
@@ -2,10 +2,6 @@
   display: block
   margin-bottom: 10px
 
-.operationList
-  display: flex
-  flex-wrap: wrap
-
   & > *
     margin: 0 10px 10px 0
 
@@ -13,7 +9,3 @@
   font-size: xx-large
   margin-bottom: 10px
   font-weight: bold
-
-.opTitle
-  font-size: x-large
-  margin-bottom: 10px
